Refactor: MarkersManager out of Booker and split Models
