+++ packages/tangle/booker.go
@@ -1,22 +1,15 @@
 package tangle
 
 import (
-	"fmt"
-	"sort"
 	"strconv"
 	"sync"
 	"time"
 
 	"github.com/cockroachdb/errors"
-	"github.com/iotaledger/hive.go/byteutils"
 	"github.com/iotaledger/hive.go/cerrors"
-	"github.com/iotaledger/hive.go/datastructure/thresholdmap"
 	"github.com/iotaledger/hive.go/datastructure/walker"
 	"github.com/iotaledger/hive.go/events"
 	"github.com/iotaledger/hive.go/identity"
-	"github.com/iotaledger/hive.go/marshalutil"
-	"github.com/iotaledger/hive.go/objectstorage"
-	"github.com/iotaledger/hive.go/stringify"
 
 	"github.com/iotaledger/goshimmer/packages/ledgerstate"
 	"github.com/iotaledger/goshimmer/packages/markers"
@@ -631,683 +624,6 @@ func markerBranchUpdatedCaller(handler interface{}, params ...interface{}) {
 
 func messageBranchUpdatedCaller(handler interface{}, params ...interface{}) {
 	handler.(func(messageID MessageID, newBranchID ledgerstate.BranchID))(params[0].(MessageID), params[1].(ledgerstate.BranchID))
-}
-
-// endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-// region MarkersManager ///////////////////////////////////////////////////////////////////////////////////////////////
-
-// MarkersManager is a Tangle component that takes care of managing the Markers which are used to infer structural
-// information about the Tangle in an efficient way.
-type MarkersManager struct {
-	tangle         *Tangle
-	discardedNodes map[identity.ID]time.Time
-	*markers.Manager
-}
-
-// NewMarkersManager is the constructor of the MarkersManager.
-func NewMarkersManager(tangle *Tangle) *MarkersManager {
-	return &MarkersManager{
-		tangle:         tangle,
-		discardedNodes: make(map[identity.ID]time.Time),
-		Manager:        markers.NewManager(tangle.Options.Store, tangle.Options.CacheTimeProvider),
-	}
-}
-
-// InheritStructureDetails returns the structure Details of a Message that are derived from the StructureDetails of its
-// strong and like parents.
-func (m *MarkersManager) InheritStructureDetails(message *Message, structureDetails []*markers.StructureDetails, sequenceAlias markers.SequenceAlias) (newStructureDetails *markers.StructureDetails, newSequenceCreated bool) {
-	newStructureDetails, newSequenceCreated = m.Manager.InheritStructureDetails(structureDetails, func(sequenceID markers.SequenceID, currentHighestIndex markers.Index) bool {
-		nodeID := identity.NewID(message.IssuerPublicKey())
-		bufferUsedRatio := float64(m.tangle.Scheduler.BufferSize()) / float64(m.tangle.Scheduler.MaxBufferSize())
-		nodeQueueRatio := float64(m.tangle.Scheduler.NodeQueueSize(nodeID)) / float64(m.tangle.Scheduler.BufferSize())
-		if bufferUsedRatio > 0.01 && nodeQueueRatio > 0.1 {
-			return false
-		}
-		if discardTime, ok := m.discardedNodes[nodeID]; ok && time.Since(discardTime) < time.Minute {
-			return false
-		} else if ok && time.Since(discardTime) >= time.Minute {
-			delete(m.discardedNodes, nodeID)
-		}
-		return m.tangle.Options.IncreaseMarkersIndexCallback(sequenceID, currentHighestIndex)
-	}, sequenceAlias)
-	if newStructureDetails.IsPastMarker {
-		m.SetMessageID(newStructureDetails.PastMarkers.Marker(), message.ID())
-		m.tangle.Utils.WalkMessageMetadata(m.propagatePastMarkerToFutureMarkers(newStructureDetails.PastMarkers.Marker()), message.ParentsByType(StrongParentType))
-	}
-
-	return
-}
-
-// MessageID retrieves the MessageID of the given Marker.
-func (m *MarkersManager) MessageID(marker *markers.Marker) (messageID MessageID) {
-	m.tangle.Storage.MarkerMessageMapping(marker).Consume(func(markerMessageMapping *MarkerMessageMapping) {
-		messageID = markerMessageMapping.MessageID()
-	})
-
-	return
-}
-
-// SetMessageID associates a MessageID with the given Marker.
-func (m *MarkersManager) SetMessageID(marker *markers.Marker, messageID MessageID) {
-	m.tangle.Storage.StoreMarkerMessageMapping(NewMarkerMessageMapping(marker, messageID))
-}
-
-// BranchID returns the BranchID that is associated with the given Marker.
-func (m *MarkersManager) BranchID(marker *markers.Marker) (branchID ledgerstate.BranchID) {
-	if marker.SequenceID() == 0 {
-		return ledgerstate.MasterBranchID
-	}
-
-	m.tangle.Storage.MarkerIndexBranchIDMapping(marker.SequenceID()).Consume(func(markerIndexBranchIDMapping *MarkerIndexBranchIDMapping) {
-		branchID = markerIndexBranchIDMapping.BranchID(marker.Index())
-	})
-
-	return
-}
-
-// ConflictBranchIDs returns the ConflictBranchIDs that are associated with the given Marker.
-func (m *MarkersManager) ConflictBranchIDs(marker *markers.Marker) (branchIDs ledgerstate.BranchIDs, err error) {
-	if branchIDs, err = m.tangle.LedgerState.ResolvePendingConflictBranchIDs(ledgerstate.NewBranchIDs(m.BranchID(marker))); err != nil {
-		err = errors.Errorf("failed to resolve ConflictBranchIDs of marker %s: %w", marker, err)
-	}
-	return
-}
-
-// SetBranchID associates a BranchID with the given Marker.
-func (m *MarkersManager) SetBranchID(marker *markers.Marker, branchID ledgerstate.BranchID) (updated bool) {
-	if floorMarker, floorBranchID, exists := m.Floor(marker); exists {
-		if floorBranchID == branchID {
-			return false
-		}
-
-		if floorMarker == marker.Index() {
-			m.UnregisterSequenceAliasMapping(markers.NewSequenceAlias(floorBranchID.Bytes()), marker.SequenceID())
-
-			m.deleteBranchIDMapping(markers.NewMarker(marker.SequenceID(), floorMarker))
-		}
-
-		m.registerSequenceAliasMappingIfLastMappedMarker(marker, branchID)
-	}
-
-	m.setBranchIDMapping(marker, branchID)
-
-	return true
-}
-
-func (m *MarkersManager) setBranchIDMapping(marker *markers.Marker, branchID ledgerstate.BranchID) bool {
-	return m.tangle.Storage.MarkerIndexBranchIDMapping(marker.SequenceID(), NewMarkerIndexBranchIDMapping).Consume(func(markerIndexBranchIDMapping *MarkerIndexBranchIDMapping) {
-		markerIndexBranchIDMapping.SetBranchID(marker.Index(), branchID)
-	})
-}
-
-func (m *MarkersManager) deleteBranchIDMapping(marker *markers.Marker) bool {
-	return m.tangle.Storage.MarkerIndexBranchIDMapping(marker.SequenceID(), NewMarkerIndexBranchIDMapping).Consume(func(markerIndexBranchIDMapping *MarkerIndexBranchIDMapping) {
-		markerIndexBranchIDMapping.DeleteBranchID(marker.Index())
-	})
-}
-
-func (m *MarkersManager) registerSequenceAliasMappingIfLastMappedMarker(marker *markers.Marker, branchID ledgerstate.BranchID) {
-	if _, _, exists := m.Ceiling(markers.NewMarker(marker.SequenceID(), marker.Index()+1)); !exists {
-		m.RegisterSequenceAliasMapping(markers.NewSequenceAlias(branchID.Bytes()), marker.SequenceID())
-	}
-}
-
-// Floor returns the largest Index that is <= the given Marker, it's BranchID and a boolean value indicating if it
-// exists.
-func (m *MarkersManager) Floor(referenceMarker *markers.Marker) (marker markers.Index, branchID ledgerstate.BranchID, exists bool) {
-	m.tangle.Storage.MarkerIndexBranchIDMapping(referenceMarker.SequenceID(), NewMarkerIndexBranchIDMapping).Consume(func(markerIndexBranchIDMapping *MarkerIndexBranchIDMapping) {
-		marker, branchID, exists = markerIndexBranchIDMapping.Floor(referenceMarker.Index())
-	})
-
-	return
-}
-
-// Ceiling returns the smallest Index that is >= the given Marker, it's BranchID and a boolean value indicating if it
-// exists.
-func (m *MarkersManager) Ceiling(referenceMarker *markers.Marker) (marker markers.Index, branchID ledgerstate.BranchID, exists bool) {
-	m.tangle.Storage.MarkerIndexBranchIDMapping(referenceMarker.SequenceID(), NewMarkerIndexBranchIDMapping).Consume(func(markerIndexBranchIDMapping *MarkerIndexBranchIDMapping) {
-		marker, branchID, exists = markerIndexBranchIDMapping.Ceiling(referenceMarker.Index())
-	})
-
-	return
-}
-
-// ForEachBranchIDMapping iterates over all BranchID mappings in the given Sequence that are bigger than the given
-// thresholdIndex. Setting the thresholdIndex to 0 will iterate over all existing mappings.
-func (m *MarkersManager) ForEachBranchIDMapping(sequenceID markers.SequenceID, thresholdIndex markers.Index, callback func(mappedMarker *markers.Marker, mappedBranchID ledgerstate.BranchID)) {
-	currentMarker := markers.NewMarker(sequenceID, thresholdIndex)
-	referencingMarkerIndexInSameSequence, mappedBranchID, exists := m.Ceiling(markers.NewMarker(currentMarker.SequenceID(), currentMarker.Index()+1))
-	for ; exists; referencingMarkerIndexInSameSequence, mappedBranchID, exists = m.Ceiling(markers.NewMarker(currentMarker.SequenceID(), currentMarker.Index()+1)) {
-		currentMarker = markers.NewMarker(currentMarker.SequenceID(), referencingMarkerIndexInSameSequence)
-		callback(currentMarker, mappedBranchID)
-	}
-}
-
-// ForEachMarkerReferencingMarker executes the callback function for each Marker of other Sequences that directly
-// reference the given Marker.
-func (m *MarkersManager) ForEachMarkerReferencingMarker(referencedMarker *markers.Marker, callback func(referencingMarker *markers.Marker)) {
-	m.Sequence(referencedMarker.SequenceID()).Consume(func(sequence *markers.Sequence) {
-		sequence.ReferencingMarkers(referencedMarker.Index()).ForEachSorted(func(referencingSequenceID markers.SequenceID, referencingIndex markers.Index) bool {
-			if referencingSequenceID == referencedMarker.SequenceID() {
-				return true
-			}
-
-			callback(markers.NewMarker(referencingSequenceID, referencingIndex))
-
-			return true
-		})
-	})
-}
-
-// propagatePastMarkerToFutureMarkers updates the FutureMarkers of the strong parents of a given message when a new
-// PastMaster was assigned.
-func (m *MarkersManager) propagatePastMarkerToFutureMarkers(pastMarkerToInherit *markers.Marker) func(messageMetadata *MessageMetadata, walker *walker.Walker) {
-	return func(messageMetadata *MessageMetadata, walker *walker.Walker) {
-		updated, inheritFurther := m.UpdateStructureDetails(messageMetadata.StructureDetails(), pastMarkerToInherit)
-		if updated {
-			messageMetadata.SetModified(true)
-		}
-		if inheritFurther {
-			m.tangle.Storage.Message(messageMetadata.ID()).Consume(func(message *Message) {
-				for _, strongParentMessageID := range message.ParentsByType(StrongParentType) {
-					walker.Push(strongParentMessageID)
-				}
-			})
-		}
-	}
-}
-
-// increaseMarkersIndexCallbackStrategy implements the default strategy for increasing marker Indexes in the Tangle.
-func increaseMarkersIndexCallbackStrategy(markers.SequenceID, markers.Index) bool {
-	return true
-}
-
-// endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-// region MarkerIndexBranchIDMapping ///////////////////////////////////////////////////////////////////////////////////
-
-// MarkerIndexBranchIDMapping is a data structure that allows to map marker Indexes to a BranchID.
-type MarkerIndexBranchIDMapping struct {
-	sequenceID   markers.SequenceID
-	mapping      *thresholdmap.ThresholdMap
-	mappingMutex sync.RWMutex
-
-	objectstorage.StorableObjectFlags
-}
-
-// NewMarkerIndexBranchIDMapping creates a new MarkerIndexBranchIDMapping for the given SequenceID.
-func NewMarkerIndexBranchIDMapping(sequenceID markers.SequenceID) (markerBranchMapping *MarkerIndexBranchIDMapping) {
-	markerBranchMapping = &MarkerIndexBranchIDMapping{
-		sequenceID: sequenceID,
-		mapping:    thresholdmap.New(thresholdmap.LowerThresholdMode, markerIndexComparator),
-	}
-
-	markerBranchMapping.SetModified()
-	markerBranchMapping.Persist()
-
-	return
-}
-
-// MarkerIndexBranchIDMappingFromBytes unmarshals a MarkerIndexBranchIDMapping from a sequence of bytes.
-func MarkerIndexBranchIDMappingFromBytes(bytes []byte) (markerIndexBranchIDMapping *MarkerIndexBranchIDMapping, consumedBytes int, err error) {
-	marshalUtil := marshalutil.New(bytes)
-	if markerIndexBranchIDMapping, err = MarkerIndexBranchIDMappingFromMarshalUtil(marshalUtil); err != nil {
-		err = errors.Errorf("failed to parse MarkerIndexBranchIDMapping from MarshalUtil: %w", err)
-		return
-	}
-	consumedBytes = marshalUtil.ReadOffset()
-
-	return
-}
-
-// MarkerIndexBranchIDMappingFromMarshalUtil unmarshals a MarkerIndexBranchIDMapping using a MarshalUtil (for easier
-// unmarshalling).
-func MarkerIndexBranchIDMappingFromMarshalUtil(marshalUtil *marshalutil.MarshalUtil) (markerIndexBranchIDMapping *MarkerIndexBranchIDMapping, err error) {
-	markerIndexBranchIDMapping = &MarkerIndexBranchIDMapping{}
-	if markerIndexBranchIDMapping.sequenceID, err = markers.SequenceIDFromMarshalUtil(marshalUtil); err != nil {
-		err = errors.Errorf("failed to parse SequenceID from MarshalUtil: %w", err)
-		return
-	}
-	mappingCount, mappingCountErr := marshalUtil.ReadUint64()
-	if mappingCountErr != nil {
-		err = errors.Errorf("failed to parse reference count (%v): %w", mappingCountErr, cerrors.ErrParseBytesFailed)
-		return
-	}
-	markerIndexBranchIDMapping.mapping = thresholdmap.New(thresholdmap.LowerThresholdMode, markerIndexComparator)
-	for j := uint64(0); j < mappingCount; j++ {
-		index, indexErr := marshalUtil.ReadUint64()
-		if indexErr != nil {
-			err = errors.Errorf("failed to parse Index (%v): %w", indexErr, cerrors.ErrParseBytesFailed)
-			return
-		}
-
-		branchID, branchIDErr := ledgerstate.BranchIDFromMarshalUtil(marshalUtil)
-		if branchIDErr != nil {
-			err = errors.Errorf("failed to parse BranchID: %w", branchIDErr)
-			return
-		}
-
-		markerIndexBranchIDMapping.mapping.Set(markers.Index(index), branchID)
-	}
-
-	return
-}
-
-// MarkerIndexBranchIDMappingFromObjectStorage restores a MarkerIndexBranchIDMapping that was stored in the object
-// storage.
-func MarkerIndexBranchIDMappingFromObjectStorage(key []byte, data []byte) (markerIndexBranchIDMapping objectstorage.StorableObject, err error) {
-	if markerIndexBranchIDMapping, _, err = MarkerIndexBranchIDMappingFromBytes(byteutils.ConcatBytes(key, data)); err != nil {
-		err = errors.Errorf("failed to parse MarkerIndexBranchIDMapping from bytes: %w", err)
-		return
-	}
-
-	return
-}
-
-// SequenceID returns the SequenceID that this MarkerIndexBranchIDMapping represents.
-func (m *MarkerIndexBranchIDMapping) SequenceID() markers.SequenceID {
-	return m.sequenceID
-}
-
-// BranchID returns the BranchID that is associated to the given marker Index.
-func (m *MarkerIndexBranchIDMapping) BranchID(markerIndex markers.Index) (branchID ledgerstate.BranchID) {
-	m.mappingMutex.RLock()
-	defer m.mappingMutex.RUnlock()
-
-	value, exists := m.mapping.Get(markerIndex)
-	if !exists {
-		panic(fmt.Sprintf("tried to retrieve the BranchID of unknown marker.%s", markerIndex))
-	}
-
-	return value.(ledgerstate.BranchID)
-}
-
-// SetBranchID creates a mapping between the given marker Index and the given BranchID.
-func (m *MarkerIndexBranchIDMapping) SetBranchID(index markers.Index, branchID ledgerstate.BranchID) {
-	m.mappingMutex.Lock()
-	defer m.mappingMutex.Unlock()
-
-	m.mapping.Set(index, branchID)
-	m.SetModified()
-}
-
-// DeleteBranchID deletes a mapping between the given marker Index and the stored BranchID.
-func (m *MarkerIndexBranchIDMapping) DeleteBranchID(index markers.Index) {
-	m.mappingMutex.Lock()
-	defer m.mappingMutex.Unlock()
-
-	m.mapping.Delete(index)
-	m.SetModified()
-}
-
-// Floor returns the largest Index that is <= the given Index which has a mapped BranchID (and a boolean value
-// indicating if it exists).
-func (m *MarkerIndexBranchIDMapping) Floor(index markers.Index) (marker markers.Index, branchID ledgerstate.BranchID, exists bool) {
-	m.mappingMutex.RLock()
-	defer m.mappingMutex.RUnlock()
-
-	if untypedIndex, untypedBranchID, exists := m.mapping.Floor(index); exists {
-		return untypedIndex.(markers.Index), untypedBranchID.(ledgerstate.BranchID), true
-	}
-
-	return 0, ledgerstate.UndefinedBranchID, false
-}
-
-// Ceiling returns the smallest Index that is >= the given Index which has a mapped BranchID (and a boolean value
-// indicating if it exists).
-func (m *MarkerIndexBranchIDMapping) Ceiling(index markers.Index) (marker markers.Index, branchID ledgerstate.BranchID, exists bool) {
-	m.mappingMutex.RLock()
-	defer m.mappingMutex.RUnlock()
-
-	if untypedIndex, untypedBranchID, exists := m.mapping.Ceiling(index); exists {
-		return untypedIndex.(markers.Index), untypedBranchID.(ledgerstate.BranchID), true
-	}
-
-	return 0, ledgerstate.UndefinedBranchID, false
-}
-
-// Bytes returns a marshaled version of the MarkerIndexBranchIDMapping.
-func (m *MarkerIndexBranchIDMapping) Bytes() []byte {
-	return byteutils.ConcatBytes(m.ObjectStorageKey(), m.ObjectStorageValue())
-}
-
-// String returns a human-readable version of the MarkerIndexBranchIDMapping.
-func (m *MarkerIndexBranchIDMapping) String() string {
-	m.mappingMutex.RLock()
-	defer m.mappingMutex.RUnlock()
-
-	indexes := make([]markers.Index, 0)
-	branchIDs := make(map[markers.Index]ledgerstate.BranchID)
-	m.mapping.ForEach(func(node *thresholdmap.Element) bool {
-		index := node.Key().(markers.Index)
-		indexes = append(indexes, index)
-		branchIDs[index] = node.Value().(ledgerstate.BranchID)
-
-		return true
-	})
-
-	sort.Slice(indexes, func(i, j int) bool {
-		return indexes[i] < indexes[j]
-	})
-
-	mapping := stringify.StructBuilder("Mapping")
-	for i, referencingIndex := range indexes {
-		thresholdStart := strconv.FormatUint(uint64(referencingIndex), 10)
-		thresholdEnd := "INF"
-		if len(indexes) > i+1 {
-			thresholdEnd = strconv.FormatUint(uint64(indexes[i+1])-1, 10)
-		}
-
-		if thresholdStart == thresholdEnd {
-			mapping.AddField(stringify.StructField("Index("+thresholdStart+")", branchIDs[referencingIndex]))
-		} else {
-			mapping.AddField(stringify.StructField("Index("+thresholdStart+" ... "+thresholdEnd+")", branchIDs[referencingIndex]))
-		}
-	}
-
-	return stringify.Struct("MarkerIndexBranchIDMapping",
-		stringify.StructField("sequenceID", m.sequenceID),
-		stringify.StructField("mapping", mapping),
-	)
-}
-
-// Update is disabled and panics if it ever gets called - it is required to match the StorableObject interface.
-func (m *MarkerIndexBranchIDMapping) Update(objectstorage.StorableObject) {
-	panic("updates disabled")
-}
-
-// ObjectStorageKey returns the key that is used to store the object in the database. It is required to match the
-// StorableObject interface.
-func (m *MarkerIndexBranchIDMapping) ObjectStorageKey() []byte {
-	return m.sequenceID.Bytes()
-}
-
-// ObjectStorageValue marshals the ConflictBranch into a sequence of bytes that are used as the value part in the
-// object storage.
-func (m *MarkerIndexBranchIDMapping) ObjectStorageValue() []byte {
-	m.mappingMutex.RLock()
-	defer m.mappingMutex.RUnlock()
-
-	marshalUtil := marshalutil.New()
-	marshalUtil.WriteUint64(uint64(m.mapping.Size()))
-	m.mapping.ForEach(func(node *thresholdmap.Element) bool {
-		marshalUtil.Write(node.Key().(markers.Index))
-		marshalUtil.Write(node.Value().(ledgerstate.BranchID))
-
-		return true
-	})
-
-	return marshalUtil.Bytes()
-}
-
-// markerIndexComparator is a comparator for marker Indexes.
-func markerIndexComparator(a, b interface{}) int {
-	aCasted := a.(markers.Index)
-	bCasted := b.(markers.Index)
-
-	switch {
-	case aCasted < bCasted:
-		return -1
-	case aCasted > bCasted:
-		return 1
-	default:
-		return 0
-	}
-}
-
-// code contract (make sure the type implements all required methods)
-var _ objectstorage.StorableObject = &MarkerIndexBranchIDMapping{}
-
-// endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-// region CachedMarkerIndexBranchIDMapping /////////////////////////////////////////////////////////////////////////////
-
-// CachedMarkerIndexBranchIDMapping is a wrapper for the generic CachedObject returned by the object storage that
-// overrides the accessor methods with a type-casted one.
-type CachedMarkerIndexBranchIDMapping struct {
-	objectstorage.CachedObject
-}
-
-// Retain marks the CachedObject to still be in use by the program.
-func (c *CachedMarkerIndexBranchIDMapping) Retain() *CachedMarkerIndexBranchIDMapping {
-	return &CachedMarkerIndexBranchIDMapping{c.CachedObject.Retain()}
-}
-
-// Unwrap is the type-casted equivalent of Get. It returns nil if the object does not exist.
-func (c *CachedMarkerIndexBranchIDMapping) Unwrap() *MarkerIndexBranchIDMapping {
-	untypedObject := c.Get()
-	if untypedObject == nil {
-		return nil
-	}
-
-	typedObject := untypedObject.(*MarkerIndexBranchIDMapping)
-	if typedObject == nil || typedObject.IsDeleted() {
-		return nil
-	}
-
-	return typedObject
-}
-
-// Consume unwraps the CachedObject and passes a type-casted version to the consumer (if the object is not empty - it
-// exists). It automatically releases the object when the consumer finishes.
-func (c *CachedMarkerIndexBranchIDMapping) Consume(consumer func(markerIndexBranchIDMapping *MarkerIndexBranchIDMapping), forceRelease ...bool) (consumed bool) {
-	return c.CachedObject.Consume(func(object objectstorage.StorableObject) {
-		consumer(object.(*MarkerIndexBranchIDMapping))
-	}, forceRelease...)
-}
-
-// String returns a human-readable version of the CachedMarkerIndexBranchIDMapping.
-func (c *CachedMarkerIndexBranchIDMapping) String() string {
-	return stringify.Struct("CachedMarkerIndexBranchIDMapping",
-		stringify.StructField("CachedObject", c.Unwrap()),
-	)
-}
-
-// endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-// region MarkerMessageMapping /////////////////////////////////////////////////////////////////////////////////////////
-
-// MarkerMessageMappingPartitionKeys defines the "layout" of the key. This enables prefix iterations in the object
-// storage.
-var MarkerMessageMappingPartitionKeys = objectstorage.PartitionKey(markers.SequenceIDLength, markers.IndexLength)
-
-// MarkerMessageMapping is a data structure that denotes a mapping from a Marker to a Message.
-type MarkerMessageMapping struct {
-	marker    *markers.Marker
-	messageID MessageID
-
-	objectstorage.StorableObjectFlags
-}
-
-// NewMarkerMessageMapping is the constructor for the MarkerMessageMapping.
-func NewMarkerMessageMapping(marker *markers.Marker, messageID MessageID) *MarkerMessageMapping {
-	return &MarkerMessageMapping{
-		marker:    marker,
-		messageID: messageID,
-	}
-}
-
-// MarkerMessageMappingFromBytes unmarshals an MarkerMessageMapping from a sequence of bytes.
-func MarkerMessageMappingFromBytes(bytes []byte) (individuallyMappedMessage *MarkerMessageMapping, consumedBytes int, err error) {
-	marshalUtil := marshalutil.New(bytes)
-	if individuallyMappedMessage, err = MarkerMessageMappingFromMarshalUtil(marshalUtil); err != nil {
-		err = errors.Errorf("failed to parse MarkerMessageMapping from MarshalUtil: %w", err)
-		return
-	}
-	consumedBytes = marshalUtil.ReadOffset()
-
-	return
-}
-
-// MarkerMessageMappingFromMarshalUtil unmarshals an MarkerMessageMapping using a MarshalUtil (for easier unmarshalling).
-func MarkerMessageMappingFromMarshalUtil(marshalUtil *marshalutil.MarshalUtil) (markerMessageMapping *MarkerMessageMapping, err error) {
-	markerMessageMapping = &MarkerMessageMapping{}
-	if markerMessageMapping.marker, err = markers.MarkerFromMarshalUtil(marshalUtil); err != nil {
-		err = errors.Errorf("failed to parse Marker from MarshalUtil: %w", err)
-		return
-	}
-	if markerMessageMapping.messageID, err = ReferenceFromMarshalUtil(marshalUtil); err != nil {
-		err = errors.Errorf("failed to parse MessageID from MarshalUtil: %w", err)
-		return
-	}
-
-	return
-}
-
-// MarkerMessageMappingFromObjectStorage is a factory method that creates a new MarkerMessageMapping instance
-// from a storage key of the object storage. It is used by the object storage, to create new instances of this entity.
-func MarkerMessageMappingFromObjectStorage(key, value []byte) (result objectstorage.StorableObject, err error) {
-	if result, _, err = MarkerMessageMappingFromBytes(byteutils.ConcatBytes(key, value)); err != nil {
-		err = errors.Errorf("failed to parse MarkerMessageMapping from bytes: %w", err)
-		return
-	}
-
-	return
-}
-
-// Marker returns the Marker that is mapped to a MessageID.
-func (m *MarkerMessageMapping) Marker() *markers.Marker {
-	return m.marker
-}
-
-// MessageID returns the MessageID of the Marker.
-func (m *MarkerMessageMapping) MessageID() MessageID {
-	return m.messageID
-}
-
-// Bytes returns a marshaled version of the MarkerMessageMapping.
-func (m *MarkerMessageMapping) Bytes() []byte {
-	return byteutils.ConcatBytes(m.ObjectStorageKey(), m.ObjectStorageValue())
-}
-
-// String returns a human-readable version of the MarkerMessageMapping.
-func (m *MarkerMessageMapping) String() string {
-	return stringify.Struct("MarkerMessageMapping",
-		stringify.StructField("marker", m.marker),
-		stringify.StructField("messageID", m.messageID),
-	)
-}
-
-// Update is disabled and panics if it ever gets called - it is required to match the StorableObject interface.
-func (m *MarkerMessageMapping) Update(objectstorage.StorableObject) {
-	panic("updates disabled")
-}
-
-// ObjectStorageKey returns the key that is used to store the object in the database. It is required to match the
-// StorableObject interface.
-func (m *MarkerMessageMapping) ObjectStorageKey() []byte {
-	return m.marker.Bytes()
-}
-
-// ObjectStorageValue marshals the MarkerMessageMapping into a sequence of bytes that are used as the value part in
-// the object storage.
-func (m *MarkerMessageMapping) ObjectStorageValue() []byte {
-	return m.messageID.Bytes()
-}
-
-// code contract (make sure the type implements all required methods).
-var _ objectstorage.StorableObject = &MarkerMessageMapping{}
-
-// endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-// region CachedMarkerMessageMapping ///////////////////////////////////////////////////////////////////////////////////
-
-// CachedMarkerMessageMapping is a wrapper for the generic CachedObject returned by the object storage that overrides
-// the accessor methods with a type-casted one.
-type CachedMarkerMessageMapping struct {
-	objectstorage.CachedObject
-}
-
-// Retain marks the CachedObject to still be in use by the program.
-func (c *CachedMarkerMessageMapping) Retain() *CachedMarkerMessageMapping {
-	return &CachedMarkerMessageMapping{c.CachedObject.Retain()}
-}
-
-// Unwrap is the type-casted equivalent of Get. It returns nil if the object does not exist.
-func (c *CachedMarkerMessageMapping) Unwrap() *MarkerMessageMapping {
-	untypedObject := c.Get()
-	if untypedObject == nil {
-		return nil
-	}
-
-	typedObject := untypedObject.(*MarkerMessageMapping)
-	if typedObject == nil || typedObject.IsDeleted() {
-		return nil
-	}
-
-	return typedObject
-}
-
-// Consume unwraps the CachedObject and passes a type-casted version to the consumer (if the object is not empty - it
-// exists). It automatically releases the object when the consumer finishes.
-func (c *CachedMarkerMessageMapping) Consume(consumer func(markerMessageMapping *MarkerMessageMapping), forceRelease ...bool) (consumed bool) {
-	return c.CachedObject.Consume(func(object objectstorage.StorableObject) {
-		consumer(object.(*MarkerMessageMapping))
-	}, forceRelease...)
-}
-
-// String returns a human-readable version of the CachedMarkerMessageMapping.
-func (c *CachedMarkerMessageMapping) String() string {
-	return stringify.Struct("CachedMarkerMessageMapping",
-		stringify.StructField("CachedObject", c.Unwrap()),
-	)
-}
-
-// endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-// region CachedMarkerMessageMappings //////////////////////////////////////////////////////////////////////////////////
-
-// CachedMarkerMessageMappings defines a slice of *CachedMarkerMessageMapping.
-type CachedMarkerMessageMappings []*CachedMarkerMessageMapping
-
-// Unwrap is the type-casted equivalent of Get. It returns a slice of unwrapped objects with the object being nil if it
-// does not exist.
-func (c CachedMarkerMessageMappings) Unwrap() (unwrappedMarkerMessageMappings []*MarkerMessageMapping) {
-	unwrappedMarkerMessageMappings = make([]*MarkerMessageMapping, len(c))
-	for i, cachedMarkerMessageMapping := range c {
-		untypedObject := cachedMarkerMessageMapping.Get()
-		if untypedObject == nil {
-			continue
-		}
-
-		typedObject := untypedObject.(*MarkerMessageMapping)
-		if typedObject == nil || typedObject.IsDeleted() {
-			continue
-		}
-
-		unwrappedMarkerMessageMappings[i] = typedObject
-	}
-
-	return
-}
-
-// Consume iterates over the CachedObjects, unwraps them and passes a type-casted version to the consumer (if the object
-// is not empty - it exists). It automatically releases the object when the consumer finishes. It returns true, if at
-// least one object was consumed.
-func (c CachedMarkerMessageMappings) Consume(consumer func(markerMessageMapping *MarkerMessageMapping), forceRelease ...bool) (consumed bool) {
-	for _, cachedMarkerMessageMapping := range c {
-		consumed = cachedMarkerMessageMapping.Consume(consumer, forceRelease...) || consumed
-	}
-
-	return
-}
-
-// Release is a utility function that allows us to release all CachedObjects in the collection.
-func (c CachedMarkerMessageMappings) Release(force ...bool) {
-	for _, cachedMarkerMessageMapping := range c {
-		cachedMarkerMessageMapping.Release(force...)
-	}
-}
-
-// String returns a human-readable version of the CachedMarkerMessageMappings.
-func (c CachedMarkerMessageMappings) String() string {
-	structBuilder := stringify.StructBuilder("CachedMarkerMessageMappings")
-	for i, cachedMarkerMessageMapping := range c {
-		structBuilder.AddField(stringify.StructField(strconv.Itoa(i), cachedMarkerMessageMapping))
-	}
-
-	return structBuilder.String()
 }
 
 // endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
